LikedReview 추가, schema refactor 및 테스트코드 수정

+++ src/main/java/com/deokhugam/domain/poweruser/entity/PowerUser.java
@@ -37,7 +37,7 @@ public class PowerUser extends BaseEntity {
     private Long commentCount = 0L;
 
     @Column(name = "review_score_sum", nullable = false)
-    private Long reviewScoreSum = 0L;
+    private Double reviewScoreSum = 0.0;
 
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
     @JoinColumn(name = "user_id", nullable = false)
